requirements-docker.txt
-e .[docker]
-r requirements.txt
appdirs==1.4.3
bonobo-docker==0.6.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
docker-pycreds==0.3.0
docker==2.7.0
fs==2.0.27
graphviz==0.8.4
idna==2.7
jinja2==2.10
markupsafe==1.0
mondrian==0.7.0
packaging==17.1
pbr==4.2.0
psutil==5.4.6
pyparsing==2.2.0
python-slugify==1.2.5
pytz==2018.5
requests==2.19.1
semantic-version==2.6.0
six==1.11.0
stevedore==1.29.0
typing==3.6.4
unidecode==1.0.22
urllib3==1.23
websocket-client==0.48.0
whistle==1.0.1
